--- a/src/azure-cli/azure/cli/command_modules/rdbms/tests/latest/test_rdbms_commands.py
+++ b/src/azure-cli/azure/cli/command_modules/rdbms/tests/latest/test_rdbms_commands.py
@@ -1026,7 +1026,7 @@ class ReplicationPostgreSqlMgmtScenarioTest(ScenarioTest):  # pylint: disable=to
             # restart server
             self.cmd('{} server restart -g {} --name {}'
                      .format(database_engine, resource_group, server), checks=NoneCheck())
-            sleep(120)
+            time.sleep(120)
 
         # test replica create
         self.cmd('{} server replica create -g {} -n {} -l westus --sku-name {} '