**Description**<!--Mandatory-->
<!--Why this PR? What is changed? What is the effect? etc. A high-quality description can accelerate the review process.-->
This PR will speed up our test.

- Our list command is not pagination, it will get all the data.
   That is why some recording files are very large.
   And in test_show_built_in_policy, we choose to check the first resource instead of check all the resource to speed up.  

- We only mock time.sleep in azure-sdk-tools.
   That's why if someone use sleep(xxx), it will not work.
   https://github.com/Azure/azure-cli/blob/30216d0268f3491b8af2aa060d90a9e1f94f9b3c/src/azure-cli-testsdk/azure/cli/testsdk/scenario_tests/patches.py#L13

| Profile | Test Case | Before | After | Total saving time |
|--------|------------|---------|-------|--------------------|
| Latest | test_show_built_in_policy | 536.57s | 2.94s |  |
| 2019-03-01 | test_show_built_in_policy | 210.69s | 3.01s |  |
| Latest | test_sql_db_restore | 305.01s | 8.70s ||
| Latest | test_sql_db_restore_deleted | 184.63s | 7.39s ||
| Latest | test_postgrsql_general_purpose_replica_mgmt | 121.76s | 4.11s ||
| Latest | test_sql_tde | 10.54s | 3.03s ||
| Latest | test_sql_ledger | 7.62s | 3.00s ||
| Total |||| 1,344.64s |

![image](https://user-images.githubusercontent.com/18628534/161017281-79181c14-c23f-49de-8eab-a79a65529962.png)



**Testing Guide**
<!--Example commands with explanations.-->

**History Notes**
<!--If your PR is not customer-facing, use {Component Name} in the PR title. Otherwise, use [Component Name] to allow our pipeline to add the title as a history note. If you need multiple history notes or would like to overwrite the note from the PR title, please fill in the following templates.-->

[Component Name 1] BREAKING CHANGE: `az command a`: Make some customer-facing breaking change
[Component Name 2] `az command b`: Add some customer-facing feature

---

This checklist is used to make sure that common guidelines for a pull request are followed.

- [ ] The PR title and description has followed the guideline in [Submitting Pull Requests](https://github.com/Azure/azure-cli/tree/dev/doc/authoring_command_modules#submitting-pull-requests).

- [ ] I adhere to the [Command Guidelines](https://github.com/Azure/azure-cli/blob/dev/doc/command_guidelines.md).

- [ ] I adhere to the [Error Handling Guidelines](https://github.com/Azure/azure-cli/blob/dev/doc/error_handling_guidelines.md).